requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
google-genai==0.3.0
python-multipart==0.0.6
pydantic==2.5.3
